updated repository with functional functions

+++ ft_atoi.c
@@ -6,15 +6,15 @@
 /*   By: nhariman <marvin@codam.nl>                   +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/10/28 15:55:47 by nhariman       #+#    #+#                */
-/*   Updated: 2019/11/07 21:45:15 by nhariman      ########   odam.nl         */
+/*   Updated: 2019/11/14 15:11:22 by nhariman      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 int	ft_atoi(const char *str)
 {
-	long	number;
-	int		negative;
-	int		i;
+	int	number;
+	int	negative;
+	int	i;
 
 	number = 0;
 	negative = 0;
